RAM Class Persistence: Remove dynamic Proxy classes from RCP cache
Dynamic Proxy names are assigned in OpenJDK according to the order
of the method Proxy.newProxyInstance() invoked from applications.
It's problem in restore run, because the order of the method
Proxy.newProxyInstance() invoked from applications may be changed
on conditions or events and the class lookup by the names of
$ProxyN {0...n} from RCP cache may be wrong result. The changes are
proposed to remove the classes along with unsafe classes from RCP
cache.

Fixes: #22841

Co-authored-by: Tobi Ajila tobi_ajila@ca.ibm.com