JITServer: create env var to perform all compilations locally
Due to technical reasons some of the optimizations are not applied
when the JVM runs in client mode. To evaluate the effect of those
suppressed optimizations as a whole, this commit introduces a
new environment variable: TR_SuppressAllRemoteCompilations
When this env var is set, a JVM running in client mode will
perform all compilations locally.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>