Add sending pid tracepoint for signals handled in java
Modify the predefinedHandlerWrapper(), which handles signals registered
from java code such as TERM, HUP, INT, to call a tracepoint with the
sending pid.

Move the mapping of signal names and values into the vm so it can be
shared by libjvm and libj9vm.

Check for non-NULL sigInfo, and non-zero pid.

Signed-off-by: Peter Shipton <Peter_Shipton@ca.ibm.com>