JITServer: Fix verbose={inlining} bug
`-Xjit:verbose={inlining}` prints to vlog the number of callees
that were inlined in a jitted method as well as the identity of
those callees. For a non-jitserver run this printing is done
in J9Compilation::compile(), method which is not executed when the
JVM is run in client mode. For a JITServer client the printing
of the inlined call graph is done in
`TR_RelocationRuntime::relocateAOTCodeAndData()`. However, that
code only prints something if there is at least one inlined call.

This commit changes the code so that the JITServer client will print
"0 methods inlined into..." if there are no inlined callees.
This behavior is similar to the behavior of a non-jitserver JVM.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>