Add external compiler option to ignore -Xtrace
When -Xtrace is specified, the compiler generates code to call out to
the appropriate hooks. For debugging purposes, it may be useful for the
compiler to ignore -Xtrace - for example, when trying understand what is
being interpreted.

This commit adds a new external option -XX:JITIgnoreXtrace that will
trigger this behaviour. This commit includes a new AOT feature flag to
ensure consistency between JVMs.

The one caveat here is that JNI compilations may still report to Xtrace.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>