Refactor handleServerMessage to process response in a try/catch
It may be possible that some front end query that the client makes in
order to respond to the server may throw an exception. To ensure
consistency with the server, this commit wraps the code that processes
the response in a try/catch block. If an exception is caught, the client
can send a message to the server to inform it that it should interrupt
the compilation, and the client will rethrow its exception and process
it appropriately.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>