Initialize fields of J9::Compilation in constructor
Some of the fields of J9::Compilation were initialized
immediately after the compilation object was created and
not in the constructor. In particular, the field
 `JITServer::ServerStream *_stream` was initialized after
the comp object was created, but it was used in the constructor
when adding a new object to the knownObjectTable, and this
caused a crash in the jitserver.
This commit rectifies the error by adding the relevant
parameters to the J9::Compilation constructor and
initializing the fields that were not initialized.

Fixes: #23053

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>