Enable CHTable Opts when Xnojit is set to allow AOT
Fixes issue:
https://github.com/eclipse-openj9/openj9/issues/17918

Previously, there was an issue that when -Xnojit was set, AOT loads were disabled. This is because -Xnotjit disabled CHTable Ops, which were needed for AOT header validation.

This PR enforces that if -Xnojit is specified, then AOT compilations are prohibited but CHTable optimizations are still enabled. Otherwise if recompilations are disabled for any other reason, we also disable CHTable ops.

This PR borrows a lot from https://github.com/eclipse-openj9/openj9/pull/18543