Enable CHTable Opts when Xnojit is set to allow AOT
Fixes issue:
https://github.com/eclipse-openj9/openj9/issues/17918

Previously, there was an issue that when -Xnojit was set, AOT loads were disabled. This is because -Xnotjit disabled CHTable Ops, which were needed for AOT header validation. This PR

removes some of the logic to disable CHTable Ops and replaces with some logic to check whether AOT loads are ever possible. If they aren't then we disable AOT loads.

This PR borrows a lot from https://github.com/eclipse-openj9/openj9/pull/18543

The exact logic we use to check if AOT loads are ever possible is as follows

If -Xshareclasses:none is NOT specified (since there is a default SCC even if you don't specify a name)
If -Xnoaot is NOT specified
If -Xaot:noload is NOT specified
then AOT loads are potentially possible.