Introduce recycleReservedRegions stub
This is an preparation step for a new implementation of
recycleReservedRegionsForVirtualLargeObjectHeap that will use Base Env
(instead of VLHGC specific env), since it will be invoked by common (non
VLHGC path - Indexable allocation).

To be able to call it from the common path, it has to be virtual in the
base class AllocationContext, which is defined in OMR.

The old variant is to be removed, but in interim OMR acceptance build
both of them exist. Some compilers are confused since the signatures are
different, but based on the name it's trying to associate them.

This explicit definition should help the compiler not to try to
associate them.

Signed-off-by: Aleksandar Micic <Aleksandar_Micic@ca.ibm.com>